Add more function to the market interface (#13)
