+++ eCommerce/Business/Service/ItemDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace eCommerce.Business.Service
+{
+    public class ItemDto
+    {
+        
+    }
+}
